moved products to app router
