+++ src/app/products/layout.tsx
@@ -1,0 +1,15 @@
+import { Suspense } from "react";
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <Suspense>
+        <body>{children}</body>
+      </Suspense>
+    </html>
+  );
+}
